Pull out some DOM manipulation functions
